Ajustes na api

+++ src/models/Transaction.ts
@@ -11,7 +11,7 @@ class Transaction {
   @Column('enum')
   type: 'income' | 'outcome';
 
-  @Column('int')
+  @Column('decimal')
   value: number;
 
   @Column('varchar')
